+++ .github/renovate.json
@@ -1,7 +1,7 @@
 {
   "$schema": "https://docs.renovatebot.com/renovate-schema.json",
   "extends": ["github>Boshen/renovate", "helpers:pinGitHubActionDigestsToSemver"],
-  "ignorePaths": ["crates/oxc_linter/fixtures/**"],
+  "ignorePaths": ["crates/oxc_linter/fixtures/**", "tasks/e2e"],
   "ignoreDeps": ["@types/vscode", "allocator-api2"],
   "packageRules": [
     {
