+++ pnpm-workspace.yaml
@@ -4,6 +4,7 @@ packages:
   - wasm/*
   - npm/*
   - editors/*
+  - tasks/e2e
   - tasks/transform_conformance
   - tasks/compat_data
   - tasks/mcp
@@ -11,7 +12,7 @@ packages:
 catalog:
   "@napi-rs/cli": 3.5.0
   "@napi-rs/wasm-runtime": 1.1.0
-  "@types/node": ^25.0.0
+  "@types/node": 24.1.0
   cross-env: ^10.1.0
   publint: 0.3.15
   rolldown: 1.0.0-beta.54
