+++ tasks/e2e/nestjs/src/app.controller.spec.ts
@@ -1,0 +1,22 @@
+import { expect, beforeAll, describe, it } from "vitest";
+import { Test, TestingModule } from "@nestjs/testing";
+import { AppController } from "./app.controller";
+import { AppService } from "./app.service";
+
+describe("AppController", () => {
+  let app: TestingModule;
+
+  beforeAll(async () => {
+    app = await Test.createTestingModule({
+      controllers: [AppController],
+      providers: [AppService],
+    }).compile();
+  });
+
+  describe("getHello", () => {
+    it('should return "Hello World!"', () => {
+      const appController = app.get(AppController);
+      expect(appController.getHello()).toBe("Hello World!");
+    });
+  });
+});
