+++ tasks/e2e/nestjs/test/app.e2e-spec.ts
@@ -1,0 +1,27 @@
+import type { INestApplication } from "@nestjs/common";
+import { Test } from "@nestjs/testing";
+import request from "supertest";
+import type { App } from "supertest/types";
+import { AppModule } from "./../src/app.module";
+import { afterAll, beforeAll, describe, it } from "vitest";
+
+describe("AppController (e2e)", () => {
+  let app: INestApplication<App>;
+
+  beforeAll(async () => {
+    const moduleFixture = await Test.createTestingModule({
+      imports: [AppModule],
+    }).compile();
+
+    app = moduleFixture.createNestApplication();
+    await app.init();
+  });
+
+  afterAll(async () => {
+    await app.close();
+  });
+
+  it("/ (GET)", () => {
+    return request(app.getHttpServer()).get("/").expect(200).expect("Hello World!");
+  });
+});
