--- tasks/e2e/pnpm-workspace.yaml
@@ -1,5 +1,0 @@
-packages:
-  - .
-  - tests/**
-
-verifyDepsBeforeRun: install
