--- tasks/e2e/tests/nestjs/package.json
@@ -1,23 +1,0 @@
-{
-  "name": "nest-typescript-starter",
-  "private": true,
-  "version": "1.0.0",
-  "description": "Nest TypeScript starter repository",
-  "license": "MIT",
-  "scripts": {
-    "test": "vitest"
-  },
-  "engines": {
-    "npm": ">=10.0.0",
-    "node": "^20.19.0 || >=22.12.0"
-  },
-  "dependencies": {
-    "@nestjs/common": "^11.0.1",
-    "@nestjs/core": "^11.0.1",
-    "@nestjs/platform-express": "^11.0.1",
-    "@oxc-project/runtime": "link:../../../../npm/runtime",
-    "reflect-metadata": "^0.2.2",
-    "@nestjs/testing": "^11.0.1",
-    "supertest": "^7"
-  }
-}
